+++ plugins/Duplicator/build.xml
@@ -1,3 +1,3 @@
-<project name="Cropper" default="jar">
+<project name="Duplicator" default="jar">
 	<import file="../javapluginbuild.xml"/>
 </project>
